requirements.txt
websockets>=13.0
coinbase-advanced-py>=1.2.0
PyJWT[crypto]>=2.8.0
cryptography>=41.0.0
python-dotenv>=1.0.0
anthropic>=0.39.0
pytest>=8.0
pytest-asyncio>=0.24
